Fix redirect on email block page and confirm email double call issue
