+++ client/src/pages/EmailConfirmationPage/index.ts
@@ -4,46 +4,24 @@ import { withRouter } from 'react-router';
 
 import { compose } from '../../utils';
 
-import {
-  confirmEmail,
-  logout,
-  resetLastAuthData,
-  resetEmailConfirmationResult,
-} from '../../redux/profile/actions';
+import { confirmEmail } from '../../redux/profile/actions';
 import { getInfo } from '../../redux/refProfile/actions';
-import { showLoginModal } from '../../redux/modal/actions';
-import { addManual as createNotification } from '../../redux/notifications/actions';
 
 import {
-  loading,
   emailConfirmationResult,
-  isAuthenticated,
   profileRefreshed,
-  user,
-  lastAuthData,
 } from '../../redux/profile/selectors';
-import { cartItems } from '../../redux/cart/selectors';
 
 import EmailConfirmationPage from './EmailConfirmationPage';
 
 const reduxConnect = connect(
   createStructuredSelector({
-    loading,
-    emailConfirmationResult,
-    isAuthenticated,
     profileRefreshed,
-    cartItems,
-    user,
-    lastAuthData,
+    emailConfirmationResult,
   }),
   {
     confirmEmail,
     getInfo,
-    showLoginModal,
-    logout,
-    resetLastAuthData,
-    createNotification,
-    resetEmailConfirmationResult,
   },
 );
